Modified picom setup
Fullscreen windows now have 100% opacity. This change was made because
terminals are transparent, and if put into fullscreen mode, you can see
the other windows through it.
The dwm systray was casting a literal shadow so I modified the code for
it to have class hints and added an exlusion rule from shadows for it.